requirements.txt
playwright==1.49.1
pytest-playwright==0.6.2
python-dotenv==1.0.1
